requirements.txt
django
pymysql
pillow
mysql
mysqlclient
cryptography
pytz
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client